refactor: split Codex provider internals
